Smart parking
The ultrasonic sensor has been setup to detect any object that's put in it's detection to measure distance from the sensor, so it can be able to detected when a car is parked and not parked. The LCD detects when the space is taken or not showing it to the user through WEB API like node red, and webex.