Add files via upload
Results gathered from testing on my vehicle, analysing how accurately the device published the updated messages when the car was parked or not. Improving the device code-base for the future could include larger-scale devices used on each parking space in a reasonably sized car park to show scalability with my device.